Treating every schema definition equally
