+++ src/routes/api/api.ts
@@ -53,41 +53,3 @@ export default class API extends BaseRoute {
 		}
 	}
 }
-
-export interface GraphQLSchemaDefinition {
-	definition?: string;
-	types?: string;
-	resolve?: { [key: string]: AnyFunction };
-}
-
-type AnyFunction = (...args: any[]) => any;
-
-export type GraphQLSchemaHandler = (
-	context: ServeContext,
-	request: IncomingMessage,
-	response: ServerResponse,
-) => GraphQLSchemaDefinition;
-
-export type GraphQLSchemaHandlerAggregator = (
-	request: IncomingMessage,
-	response: ServerResponse,
-	context: ServeContext,
-) => Promise<GraphQLSchemaDefinition[]>;
-
-export function isGraphQLSchemaDefinition(
-	o: any,
-): o is GraphQLSchemaDefinition {
-	if (typeof o !== "object") return false;
-
-	if (!o.types) return false;
-
-	if (o.definition) if (typeof o.definition !== "string") return false;
-
-	if (o.types) if (typeof o.types !== "string") return false;
-
-	if (o.resolve) if (typeof o.resolve !== "function") return false;
-
-	if (o.resolvers) if (!Array.isArray(o.resolvers)) return false;
-
-	return true;
-}
